requirements.txt
# keystone dependencies
pbr>=0.6,<1.0
pam>=0.1.4
WebOb>=1.2.3
eventlet>=0.13.0
greenlet>=0.3.2
netaddr>=0.7.6
PasteDeploy>=1.5.0
Paste
Routes>=1.12.3
six>=1.4.1
SQLAlchemy>=0.7.8,<=0.8.99
sqlalchemy-migrate>=0.8.2
passlib
lxml>=2.3
iso8601>=0.1.8
python-keystoneclient>=0.5.0
oslo.config>=1.2.0
Babel>=1.3
oauthlib>=0.6
dogpile.cache>=0.5.0
jsonschema>=2.0.0,<3.0.0
pecan>=0.2.0

# KDS exclusive dependencies

pycrypto>=2.6
WSME>=0.5b6
